requirements.txt
Django
django-filter
djangorestframework
Markdown
Pillow
psycopg2-binary
gunicorn